+++ Euclid/Sources/Mesh.swift
@@ -121,6 +121,11 @@ public extension Mesh {
         polygons.hasTexcoords
     }
 
+    /// A Boolean value that indicates whether the mesh includes vertex normals that differ from the face normal.
+    var hasVertexNormals: Bool {
+        polygons.hasVertexNormals
+    }
+
     /// A Boolean value that indicates whether the mesh includes vertex colors.
     var hasVertexColors: Bool {
         polygons.hasVertexColors
@@ -326,6 +331,17 @@ public extension Mesh {
         )
     }
 
+    /// Return a copy of the mesh without texture coordinates
+    func withoutTexcoords() -> Mesh {
+        Mesh(
+            unchecked: polygons.withoutTexcoords(),
+            bounds: boundsIfSet,
+            isConvex: isKnownConvex,
+            isWatertight: watertightIfSet,
+            submeshes: submeshesIfEmpty
+        )
+    }
+
     /// Returns a Boolean value that indicates if the specified point is inside the mesh.
     /// - Parameter point: The point to compare.
     /// - Returns: `true` if the point lies inside the mesh, and `false` otherwise.
@@ -350,7 +366,7 @@ public extension Mesh {
     }
 }
 
-internal extension Mesh {
+extension Mesh {
     init(
         unchecked polygons: [Polygon],
         bounds: Bounds?,