+++ Euclid/Sources/Path.swift
@@ -1,5 +1,5 @@
 //
-//  Paths.swift
+//  Path.swift
 //  Euclid
 //
 //  Created by Nick Lockwood on 03/07/2018.
@@ -216,7 +216,12 @@ public extension Path {
         var startIndex = 0
         var paths = [Path]()
         for i in subpathIndices {
-            let points = self.points[startIndex ... i]
+            var points = self.points[startIndex ... i]
+            if paths.last?.isClosed ?? false, points.count > 2,
+               points[startIndex + 1].position == points.last?.position
+            {
+                points.removeFirst()
+            }
             startIndex = i
             guard points.count > 1 else {
                 continue
@@ -436,7 +441,7 @@ public extension Polygon {
     }
 }
 
-internal extension Path {
+extension Path {
     init(unchecked points: [PathPoint], plane: Plane?, subpathIndices: [Int]?) {
         self.points = points
         self.isClosed = pointsAreClosed(unchecked: points)