+++ Euclid/Sources/Vertex.swift
@@ -140,7 +140,7 @@ public extension Vertex {
     }
 }
 
-internal extension Vertex {
+extension Vertex {
     init(
         unchecked position: Vector,
         _ normal: Vector?,
@@ -215,6 +215,13 @@ internal extension Vertex {
         return vertex
     }
 
+    /// Creates a copy of the vertex with the specified texture coordinate.
+    func with(texcoord: Vector) -> Vertex {
+        var vertex = self
+        vertex.texcoord = texcoord
+        return vertex
+    }
+
     /// Creates a copy of the vertex with the specified position.
     func with(position: Vector) -> Vertex {
         var vertex = self
@@ -231,7 +238,11 @@ internal extension Vertex {
     }
 }
 
-internal extension Collection where Element == Vertex {
+extension Collection where Element == Vertex {
+    func withoutTexcoords() -> [Vertex] {
+        map { $0.with(texcoord: .zero) }
+    }
+
     func inverted() -> [Vertex] {
         reversed().map { $0.inverted() }
     }