+++ Euclid/Tests/MeshTests.swift
@@ -329,98 +329,17 @@ class MeshTests: XCTestCase {
         }
     }
 
-    // MARK: export
+    // MARK: Optimization
 
-    func testCubeSTL() {
-        let cube = Mesh.cube().translated(by: Vector(0.5, 0.5, 0.5))
-        let stl = cube.stlString(name: "Foo")
-        XCTAssertEqual(stl, """
-        solid Foo
-        facet normal 1 0 0
-        \touter loop
-        \t\t1 0 1
-        \t\t1 0 0
-        \t\t1 1 0
-        \tendloop
-        endfacet
-        facet normal 1 0 0
-        \touter loop
-        \t\t1 0 1
-        \t\t1 1 0
-        \t\t1 1 1
-        \tendloop
-        endfacet
-        facet normal -1 0 0
-        \touter loop
-        \t\t0 0 0
-        \t\t0 0 1
-        \t\t0 1 1
-        \tendloop
-        endfacet
-        facet normal -1 0 0
-        \touter loop
-        \t\t0 0 0
-        \t\t0 1 1
-        \t\t0 1 0
-        \tendloop
-        endfacet
-        facet normal 0 1 0
-        \touter loop
-        \t\t0 1 1
-        \t\t1 1 1
-        \t\t1 1 0
-        \tendloop
-        endfacet
-        facet normal 0 1 0
-        \touter loop
-        \t\t0 1 1
-        \t\t1 1 0
-        \t\t0 1 0
-        \tendloop
-        endfacet
-        facet normal 0 -1 0
-        \touter loop
-        \t\t0 0 0
-        \t\t1 0 0
-        \t\t1 0 1
-        \tendloop
-        endfacet
-        facet normal 0 -1 0
-        \touter loop
-        \t\t0 0 0
-        \t\t1 0 1
-        \t\t0 0 1
-        \tendloop
-        endfacet
-        facet normal 0 0 1
-        \touter loop
-        \t\t0 0 1
-        \t\t1 0 1
-        \t\t1 1 1
-        \tendloop
-        endfacet
-        facet normal 0 0 1
-        \touter loop
-        \t\t0 0 1
-        \t\t1 1 1
-        \t\t0 1 1
-        \tendloop
-        endfacet
-        facet normal 0 0 -1
-        \touter loop
-        \t\t1 0 0
-        \t\t0 0 0
-        \t\t0 1 0
-        \tendloop
-        endfacet
-        facet normal 0 0 -1
-        \touter loop
-        \t\t1 0 0
-        \t\t0 1 0
-        \t\t1 1 0
-        \tendloop
-        endfacet
-        endsolid Foo
-        """)
+    func testMeshWithoutTexcoords() {
+        let mesh = Mesh.cube().withoutTexcoords()
+        XCTAssertFalse(mesh.hasTexcoords)
+    }
+
+    func testMeshWithoutVertexNormals() {
+        let cube = Mesh.cube()
+        XCTAssertFalse(cube.hasVertexNormals)
+        let sphere = Mesh.sphere().smoothNormals(.zero)
+        XCTAssertFalse(sphere.hasVertexNormals)
     }
 }