+++ Package.swift
@@ -15,7 +15,7 @@ let package = Package(
     dependencies: [
         .package(
             url: "https://github.com/nicklockwood/Euclid.git",
-            .upToNextMinor(from: "0.6.15")
+            .upToNextMinor(from: "0.6.18")
         ),
         .package(
             url: "https://github.com/nicklockwood/LRUCache.git",